backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
celery==5.3.4
redis==5.0.1
sqlalchemy==2.0.23
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
requests==2.31.0
pillow==10.1.0
websocket-client==1.6.4
PyYAML==6.0.1
psutil==5.9.6
pymysql==1.1.0
pydantic==2.5.0
aiohttp==3.9.1
websockets==12.0
asyncio-mqtt==0.16.1